phase-d(bun_runtime): RouteBundle.rs Strong->StrongOptional + add SourceMapStore::unref

+++ src/runtime/image/backend_coregraphics.rs
@@ -147,7 +147,7 @@ pub fn encode(
             width,
             height,
             fmt,
-            opts.quality,
+            i32::from(opts.quality),
             core::ptr::null_mut(),
             &mut len,
         )
@@ -165,7 +165,7 @@ pub fn encode(
             width,
             height,
             fmt,
-            opts.quality,
+            i32::from(opts.quality),
             out.as_mut_ptr(),
             &mut len,
         )
